% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimm_cv.R
\name{bimm_cv}
\alias{bimm_cv}
\title{Cross-validation for BiMMs}
\usage{
bimm_cv(
  data,
  formula,
  verbose,
  n_fold = 10,
  n_iteration = 5,
  evaluation_metric = "auc"
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{formula}{An object of class "formula" that is akin to \code{formula} in
\code{\link[lme4:lmer]{lme4::lmer()}} where both fixed and random effects are written.
Random-effects terms are distinguished by vertical bars (|) inside of
parentheses. See examples.}

\item{verbose}{An integer value of 0, 1, or 2.
\itemize{
\item If verbose = 0, nothing is printed to console.
\item If verbose = 1, select steps are printed to console.
\item If verbose = 2, almost all processes are printed to console.
}}

\item{n_fold}{an integer value of 2 or higher indicating how many folds
should be created to run cross-validation. More folds => more computation
time. Less folds => higher variability in performance assessments.}

\item{n_iteration}{A positive integer indicating how many iterations
to complete. More details on 'iterations' can be found in the 'Details'
section.}

\item{evaluation_metric}{A character value indicating which evaluation
metric to track as iterations are completed. Valid options are 'auc'
for the area underneath the receiver operator characteristic curve
and 'bri' for the Brier score.}
}
\value{
a dataframe with estimated evaluation metrics at each iteration
of the bimm fitting procedure.
}
\description{
Cross-validation for BiMMs
}
