% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital}
\alias{hospital}
\title{Hospital, Doctor, Patient (HDP) dataset.}
\format{
An object of class \code{data.frame} with 8469 rows and 22 columns.
}
\source{
\url{https://stats.idre.ucla.edu/r/codefragments/mesimulation/}
}
\usage{
hospital
}
\description{
This dataset has a three-level, hierarchical structure with patients
nested within doctors within hospitals. The purpose of this simulation
is not to do a simulation study, but rather to create a rich dataset
that can be used to show a variety of analytic techniques.
However, the principles used in this simulation could be easily
abstracted and a handful or parameters varied to do a simulation study.
}
\keyword{datasets}
